Refactored InstallOnReleaseChange
